customer-support-qa/requirements.txt
deepgram-sdk>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
requests>=2.31.0
